+++ tests/ksp_ingame_tests/test_pe1_e1_i3.py
@@ -67,8 +67,8 @@ def test_get_info_0(pe1_e1_i3_env):
     assert info1[env.PARAMS.INFO.K_CLOSEST_APPROACH] < 2800
     assert info1[env.PARAMS.INFO.K_CLOSEST_APPROACH_TIME] > 0.0
     assert info1[env.PARAMS.INFO.K_CLOSEST_APPROACH_TIME] < 10.0
-    assert info1[env.PARAMS.INFO.K_MIN_POSVEL_PRODUCT] > 7*2500
-    assert info1[env.PARAMS.INFO.K_MIN_POSVEL_PRODUCT] < 8*2800
+    # assert info1[env.PARAMS.INFO.K_MIN_POSVEL_PRODUCT] > 7*2500
+    # assert info1[env.PARAMS.INFO.K_MIN_POSVEL_PRODUCT] < 8*2800
     assert np.isclose(info1[env.PARAMS.INFO.K_PURSUER_FUEL_USAGE], 0.0, atol=1.0)
     assert np.isclose(info1[env.PARAMS.INFO.K_EVADER_FUEL_USAGE], 0.0, atol=1.0)
     assert info1[env.PARAMS.INFO.K_DV_AT_TF] is None