+++ CHANGELOG.md
@@ -9,11 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Single-value scoring function for pursuit-evade (pe1) scenarios
+
 ### Fixed
 
 ### Changed
 
 - Upgrading kRPC dependency from v0.4.8 to v0.5.2 which fixed several issues that had required hard-pinning some dependencies (setuptools, protobuf). By un-pinning these, the library is more flexible / less brittle
+- pe1 scenario reward function is now negative of scoring function and only assessed at termination (zero all other times)
 
 ### Removed